requirements.txt
pandas
yfinance
finnhub-python
transformers 
torch
python-dotenv
streamlit
firebase_admin
dotenv
langchain
transformers
sentence_transformers
accelerate
fuzzywuzzy[speedup]
langchain_community
google-cloud-firestore